refactor: relocate data-collection package to dimos/manipulation/data_collection
Moves the recorder, episode boundary, and the Piper visualization wiring out
of `dimos/visualization/rerun/` and `dimos/teleop/quest/` into one package
named for what it does. `data_collection_vis.py` is renamed to
`piper_blueprint_config.py` to make it clear it's blueprint-config code, not
a Module. Updates import paths in the quest blueprint, the recorder
import-graph guardrail test, and the auto-generated `all_blueprints.py`
registry.
