+++ dimos/robot/all_blueprints.py
@@ -132,6 +132,7 @@ all_modules = {
     "drone-tracking-module": "dimos.robot.drone.drone_tracking_module.DroneTrackingModule",
     "embedding-memory": "dimos.memory.embedding.EmbeddingMemory",
     "emitter-module": "dimos.utils.demo_image_encoding.EmitterModule",
+    "episode-boundary": "dimos.manipulation.data_collection.episode_boundary.EpisodeBoundary",
     "far-planner": "dimos.navigation.nav_stack.modules.far_planner.far_planner.FarPlanner",
     "fast-lio2": "dimos.hardware.sensors.lidar.fastlio2.module.FastLio2",
     "foxglove-bridge": "dimos.robot.foxglove_bridge.FoxgloveBridge",
@@ -180,8 +181,6 @@ all_modules = {
     "pgo": "dimos.navigation.nav_stack.modules.pgo.pgo.PGO",
     "phone-teleop-module": "dimos.teleop.phone.phone_teleop_module.PhoneTeleopModule",
     "pick-and-place-module": "dimos.manipulation.pick_and_place_module.PickAndPlaceModule",
-    "rerun-data-recorder": "dimos.visualization.rerun.recorder.RerunDataRecorder",
-    "episode-boundary": "dimos.teleop.quest.episode_boundary.EpisodeBoundary",
     "quest-teleop-module": "dimos.teleop.quest.quest_teleop_module.QuestTeleopModule",
     "real-sense-camera": "dimos.hardware.sensors.camera.realsense.camera.RealSenseCamera",
     "receiver-module": "dimos.utils.demo_image_encoding.ReceiverModule",
@@ -189,6 +188,7 @@ all_modules = {
     "reid-module": "dimos.perception.detection.reid.module.ReidModule",
     "replanning-a-star-planner": "dimos.navigation.replanning_a_star.module.ReplanningAStarPlanner",
     "rerun-bridge-module": "dimos.visualization.rerun.bridge.RerunBridgeModule",
+    "rerun-data-recorder": "dimos.manipulation.data_collection.recorder.RerunDataRecorder",
     "rerun-web-socket-server": "dimos.visualization.rerun.websocket_server.RerunWebSocketServer",
     "security-module": "dimos.experimental.security_demo.security_module.SecurityModule",
     "semantic-search": "dimos.memory2.module.SemanticSearch",
