+++ dimos/teleop/quest/blueprints.py
@@ -35,11 +35,13 @@ from dimos.manipulation.manipulation_module import ManipulationModule
 from dimos.msgs.geometry_msgs.PoseStamped import PoseStamped
 from dimos.msgs.sensor_msgs.Image import Image
 from dimos.msgs.sensor_msgs.JointState import JointState
-from dimos.teleop.quest.data_collection_vis import piper_data_collection_rerun_config
-from dimos.teleop.quest.episode_boundary import EpisodeBoundary
+from dimos.manipulation.data_collection.episode_boundary import EpisodeBoundary
+from dimos.manipulation.data_collection.piper_blueprint_config import (
+    piper_data_collection_rerun_config,
+)
+from dimos.manipulation.data_collection.recorder import RerunDataRecorder
 from dimos.teleop.quest.quest_extensions import ArmTeleopModule
 from dimos.teleop.quest.quest_types import Buttons
-from dimos.visualization.rerun.recorder import RerunDataRecorder
 from dimos.visualization.vis_module import vis_module
 
 # Arm teleop with press-and-hold engage (has rerun viz)
